+++ src/gamspy/__init__.py
@@ -22,8 +22,20 @@ from gamspy._backend.engine import EngineClient
 from gamspy._backend.neos import NeosClient
 from gamspy._config import _set_default_options, get_option, set_options
 from gamspy._container import Container
-from gamspy._model import Model, ModelStatus, Problem, Sense, SolveStatus
-from gamspy._options import FreezeOptions, ModelInstanceOptions, Options
+from gamspy._model import (
+    FileFormat,
+    Model,
+    ModelStatus,
+    Problem,
+    Sense,
+    SolveStatus,
+)
+from gamspy._options import (
+    ConvertOptions,
+    FreezeOptions,
+    ModelInstanceOptions,
+    Options,
+)
 from gamspy._serialization import deserialize, serialize
 from gamspy._symbols import (
     Alias,
@@ -69,6 +81,8 @@ __all__ = [
     "Options",
     "FreezeOptions",
     "ModelInstanceOptions",
+    "ConvertOptions",
+    "FileFormat",
     "Expression",
     "EngineClient",
     "NeosClient",